% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkAndHandleBoolVec_methods.R
\name{.checkAndHandleBoolVec}
\alias{.checkAndHandleBoolVec}
\alias{.checkAndHandleBoolVec.data.frame}
\title{.checkAndHandleBoolVec}
\usage{
.checkAndHandleBoolVec(
  x,
  by_par,
  sel,
  by_val = NULL,
  no_match = "warn",
  dups = "warn",
  dups_strict = TRUE,
  verbose = TRUE,
  ...
)

\method{.checkAndHandleBoolVec}{data.frame}(
  x,
  by_par,
  sel,
  by_val = NULL,
  no_match = "warn",
  dups = "warn",
  dups_strict = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{x}{table-like object, currently supported \code{data.frame}}

\item{by_par}{character representing colname(x), on which argument \code{sel} represents the matching}

\item{sel}{logical vector with identical(lenght(sel), nrow(x))}

\item{by_val}{atomic vector, optional, values which have been checked for matches,
just for reporting in case of non-matching cases}

\item{no_match}{character, indicating the handling of non-matching cases,
supported are 'allow' (no checks), 'msg', 'warn' and 'error'}

\item{dups}{character, indicating the handling of multiple (dups_strict==F) / duplicated matches,
supported are 'allow' (no checks), 'msg', 'warn', 'tofalse', 'first' and 'error'}

\item{dups_strict}{logical, should the handling be applied to multiple (FALSE) or duplicated (TRUE, default) matches}

\item{verbose}{logical, should the output contain detailed information about affected values (default)}

\item{...}{further arguments for methods}
}
\value{
logical vector with length of length(sel), identical to sel or modified in case of dups = 'tofalse' or 'first'
}
\description{
Wrapper to call '.swtichActions' functions
}
