% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkAndHandleBoolVec_methods.R
\name{.switchDuplActions}
\alias{.switchDuplActions}
\title{.switchDuplActions}
\usage{
.switchDuplActions(x, sel, by_par, dups, dups_strict = TRUE, verbose = TRUE)
}
\arguments{
\item{x}{data.frame}

\item{sel}{logical vector with identical(lenght(sel), nrow(x))}

\item{by_par}{character representing colname(x), on which argument \code{sel} represents the matching}

\item{dups}{character, indicating the handling of multiple (dups_strict==F) / duplicated matches,
supported are 'allow' (no checks), 'msg', 'warn', 'tofalse', 'first' and 'error'}

\item{dups_strict}{logical, should the handling be applied to multiple (FALSE) or duplicated (TRUE, default) matches}

\item{verbose}{logical, should the output contain detailed information about affected values (default)}
}
\value{
logical vector with length of length(sel), identical to sel or modified in case of 'tofalse' and 'first'
}
\description{
Helper function to distinguish between actions about handling of multiple / duplicated matches
indicated by argument \code{dups}.
}
\keyword{internal}
