% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkAndHandleBoolVec_methods.R
\name{.switchNonMatchungActions}
\alias{.switchNonMatchungActions}
\title{.switchNonMatchungActions}
\usage{
.switchNonMatchungActions(sel, by_par, no_match, by_val = NULL)
}
\arguments{
\item{sel}{logical FALSE vector with identical(lenght(sel), nrow(x))}

\item{by_par}{character representing colname(x), on which argument \code{sel} represents the matching}

\item{no_match}{character, indicating the handling of non-matching cases,
supported are 'allow' (no checks), 'msg', 'warn' and 'error'}

\item{by_val}{atomic vector, optional, values which have been checked for matches,
just for reporting in case of non-matching cases}
}
\value{
logical vector \code{sel}
}
\description{
Helper function to distinguish between actions about handling of totally non-matching cases
indicated by argument \code{no_match}.
}
\keyword{internal}
